[api] An API on Alloy
Currently there is no clean instance based API on Alloy. To compile and use modules, solver, and get instances, one needs to refer to lots of classes.

These projects provide an abstract API and an implementation of this API based on the current version of Alloy.

this is not perfect yet since we do not abstract the AST yet. That is the next step